Make builder inner class
