+++ vanilla-tweaks-bukkit/src/main/java/me/machinemaker/vanillatweaks/cloud/cooldown/CommandCooldownImpl.java
@@ -20,6 +20,8 @@
 package me.machinemaker.vanillatweaks.cloud.cooldown;
 
 import cloud.commandframework.keys.CloudKey;
+import cloud.commandframework.keys.SimpleCloudKey;
+import java.util.UUID;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.framework.qual.DefaultQualifier;
@@ -27,8 +29,44 @@ import org.checkerframework.framework.qual.DefaultQualifier;
 @DefaultQualifier(NonNull.class)
 record CommandCooldownImpl<C>(
         CloudKey<Void> key,
-        CooldownDuration<C> cooldownDuration,
+        CommandCooldownDuration<C> cooldownDuration,
         @Nullable CommandCooldownNotifier<C> notifier
 ) implements CommandCooldown<C> {
 
+    @DefaultQualifier(NonNull.class)
+    static final class BuilderImpl<C> implements Builder<C> {
+
+        private final CommandCooldownDuration<C> cooldownDuration;
+        private @Nullable CommandCooldownNotifier<C> notifier;
+        private @Nullable CloudKey<Void> cooldownKey;
+
+        BuilderImpl(final CommandCooldownDuration<C> cooldownDuration) {
+            this.cooldownDuration = cooldownDuration;
+        }
+
+        @Override
+        public Builder<C> notifier(final CommandCooldownNotifier<C> notifier) {
+            this.notifier = notifier;
+            return this;
+        }
+
+        @Override
+        public Builder<C> key(final CloudKey<Void> cooldownKey) {
+            this.cooldownKey = cooldownKey;
+            return this;
+        }
+
+        private CloudKey<Void> getOrCreateKey() {
+            @Nullable CloudKey<Void> key = this.cooldownKey;
+            if (key == null) {
+                key = SimpleCloudKey.of(UUID.randomUUID().toString());
+            }
+            return key;
+        }
+
+        @Override
+        public CommandCooldown<C> build() {
+            return new CommandCooldownImpl<>(this.getOrCreateKey(), this.cooldownDuration, this.notifier);
+        }
+    }
 }
